Add nvidia_tegra_drivers_support interface
This interface will give snaps access to the hardware on nvidia tegra
platforms. It includes the udev rules contained in the
nvidia-tegra-drivers-36 .deb package in the ubuntu-tegra/updates ppa[1].

The apparmor profiles and seccomp rules were determined iteratively
using snappy-debug and testing it with three different snaps that were
created. One for the nvidia-tegra-drivers-36 package which also contains
the `nvidia-smi` tool, one for cuda-samples[2] and one for
libnvinfer-samples[3] in nvidia's jetson/common package repository which
contains TensorRT samples.

[1] https://launchpad.net/~ubuntu-tegra/+archive/ubuntu/updates
[2] https://github.com/NVIDIA/cuda-samples
[3] https://repo.download.nvidia.com/jetson#Jetpack%206.0

Signed-off-by: Sebastian Weyer <sebastian.weyer@canonical.com>